refactor: Moved validation to services
