+++ src/main/java/faang/school/projectservice/service/ProjectService.java
@@ -1,9 +1,12 @@
 package faang.school.projectservice.service;
 
+import faang.school.projectservice.dto.vacancy.OpenVacancyRequestDto;
 import faang.school.projectservice.model.Project;
 
 import java.util.Optional;
 
 public interface ProjectService {
     Optional<Project> getProjectByIdOrEmpty(long projectId);
+
+    Project validateAndGetProject(OpenVacancyRequestDto requestDto);
 }
