+++ src/main/java/faang/school/projectservice/service/ProjectServiceImpl.java
@@ -1,5 +1,7 @@
 package faang.school.projectservice.service;
 
+import faang.school.projectservice.dto.vacancy.OpenVacancyRequestDto;
+import faang.school.projectservice.exception.DataValidationException;
 import faang.school.projectservice.model.Project;
 import faang.school.projectservice.repository.ProjectRepository;
 import lombok.RequiredArgsConstructor;
@@ -15,5 +17,11 @@ public class ProjectServiceImpl implements ProjectService {
 
     public Optional<Project> getProjectByIdOrEmpty(long projectId) {
         return projectRepository.findById(projectId);
+    }
+
+    public Project validateAndGetProject(OpenVacancyRequestDto requestDto) {
+        return getProjectByIdOrEmpty(requestDto.projectId())
+                .orElseThrow(() -> new DataValidationException(
+                        "Project with id %d is not found".formatted(requestDto.projectId())));
     }
 }
