--- src/main/java/faang/school/projectservice/validator/OpenVacancyRequestValidator.java
@@ -1,48 +1,0 @@
-package faang.school.projectservice.validator;
-
-import faang.school.projectservice.dto.vacancy.OpenVacancyRequestDto;
-import faang.school.projectservice.exception.DataValidationException;
-import faang.school.projectservice.model.Project;
-import faang.school.projectservice.model.TeamMember;
-import faang.school.projectservice.model.TeamRole;
-import faang.school.projectservice.service.ProjectService;
-import faang.school.projectservice.service.TeamMemberService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-
-@Component
-@RequiredArgsConstructor
-public class OpenVacancyRequestValidator {
-
-    private final ProjectService projectService;
-    private final TeamMemberService teamMemberService;
-
-    public Project validateAndGetProject(OpenVacancyRequestDto requestDto) {
-        return projectService.getProjectByIdOrEmpty(requestDto.projectId())
-                .orElseThrow(() -> new DataValidationException(
-                        "Project with id %d is not found".formatted(requestDto.projectId())));
-    }
-
-    public TeamMember validateAndGetAuthor(OpenVacancyRequestDto requestDto) {
-        var author = teamMemberService.getTeamMemberById(requestDto.authorId())
-                .orElseThrow(() -> new DataValidationException(
-                        "Author with id %d is not found".formatted(requestDto.authorId())));
-
-        var isRightRole = author.getRoles()
-                .stream()
-                .anyMatch(role -> role == TeamRole.OWNER || role == TeamRole.MANAGER);
-        if (!isRightRole) {
-            throw new DataValidationException(
-                    "Current author roles are %s. Only OWNER and MANAGER is possible".formatted(
-                            String.join(",", author.getRoles().stream().map(Enum::toString).toList())));
-        }
-
-        return author;
-    }
-
-    public void validateSalary(OpenVacancyRequestDto requestDto) {
-        if (requestDto.salary() != null && requestDto.salary() <= 0) {
-            throw new DataValidationException("Negative or null salary is crazy");
-        }
-    }
-}
